詳細画面と一覧画面の分離
