+++ src/app/(default)/dog/[dogId]/components/DogDetail/DogDetail.module.scss
