+++ src/app/(default)/dog/[dogId]/components/DogDetail/index.tsx
